mod(codec): move `osm` into separate module to isolate import
