+++ libs/routers_codec/src/lib.rs
@@ -6,43 +6,4 @@ use mimalloc::MiMalloc;
 #[cfg(feature = "mimalloc")]
 static GLOBAL: MiMalloc = MiMalloc;
 
-// Exposed modules
-pub mod blob;
-pub mod block;
-pub mod element;
-
-// Hidden modules
-#[doc(hidden)]
-pub mod error;
-#[doc(hidden)]
-pub mod parallel;
-#[doc(hidden)]
-pub mod test;
-
-// Inlined structs
-#[doc(inline)]
-pub use blob::iterator::BlobIterator;
-#[doc(inline)]
-pub use block::iterator::BlockIterator;
-#[doc(inline)]
-pub use element::iterator::ElementIterator;
-#[doc(inline)]
-pub use element::processed_iterator::ProcessedElementIterator;
-
-// Doc-Linking
-#[doc(hidden)]
-pub use blob::item::BlobItem;
-#[doc(hidden)]
-pub use block::item::BlockItem;
-#[doc(hidden)]
-pub use element::item::Element;
-#[doc(hidden)]
-pub use osm::*;
-#[doc(inline)]
-pub use parallel::Parallel;
-
-// Protocol Buffer Includes
-pub mod osm {
-    //! OpenStreetMaps Protobuf Definitions
-    include!(concat!(env!("OUT_DIR"), "/osmpbf.rs"));
-}
+pub mod osm;
