+++ libs/routers_codec/src/osm/blob/item.rs
