+++ libs/routers_codec/src/osm/blob/mod.rs
