+++ libs/routers_codec/src/osm/parallel.rs
