+++ src/route/transition/solver/methods.rs
@@ -1,5 +1,5 @@
 use crate::route::transition::*;
-use codec::element::variants::OsmEntryId;
+use codec::osm::element::variants::OsmEntryId;
 
 #[derive(Debug, Default, Copy, Clone)]
 pub enum ResolutionMethod {
